requirements.txt
openai
transformers
sentence-transformers
faiss-cpu
pandas
scikit-learn
streamlit
python-dotenv
